% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{diff_deseq2_count}
\alias{diff_deseq2_count}
\title{Do differential expression analysis on read count data using \code{DESeq2}}
\usage{
diff_deseq2_count(count_mat, label_list, groups)
}
\arguments{
\item{count_mat}{A matrix-like data object containing the read count data, with rows corresponding to genes and columns to samples.}

\item{label_list}{a dataframe contain samples in first column and groups information
in second column in the form of factors.}

\item{groups}{the same as \code{contrast} argument in \code{\link[DESeq2]{results}}, as the name of a factor in the design formula is \code{condition}}
}
\value{
A dataframe with a row for the number top genes and the following columns:
\item{baseMean}{}
\item{log2FoldChange}{log2-fold change of expression between conditions being tested.}
\item{lfcSE}{the standard error of the \code{log2FoldChange}}
\item{stat}{Wald statistic}
\item{pvalue}{p-values.}
\item{padj}{adjusted p-values by 'BH'}
}
\description{
Do differential expression analysis using \code{DESeq2} package,
this is for the read count data in RNA-seq.
}
\examples{
data(OSCC)
res <- diff_deseq2_count(count_mat = OSCC$count, label_list = OSCC$group,
                         groups = c("condition", "tumor", "normal"))
}
\author{
Erjie Zhao \href{mailto:2055469819@qq.com}{2055469819@qq.com}
}
