% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR.R
\name{diff_edger_count}
\alias{diff_edger_count}
\title{Do differential expression analysis on read count data using \code{edgeR}}
\usage{
diff_edger_count(count_mat, label_list, groups)
}
\arguments{
\item{count_mat}{A matrix-like data object containing the read count data, with rows corresponding to genes and columns to samples.}

\item{label_list}{a dataframe contain samples in first column and groups information
in second column in the form of factors.}

\item{groups}{expressions, or character strings which can be parsed to expressions,
specifying contrasts, it corresponds to \code{\link[limma]{makeContrasts}}.}
}
\value{
A dataframe with a row for the number top genes and the following columns:
\item{logFC}{log2-fold change of expression between conditions being tested.}
\item{logCPM}{average log2-counts per million.}
\item{LR}{likelihood ratio statistics.}
\item{PValue}{p-values.}
\item{FDR}{false discovery rate.}
}
\description{
Do differential expression analysis using \code{edgeR} package,
this is for the read count data in RNA-seq.
}
\examples{
data(OSCC)
res <- diff_edger_count(count_mat = OSCC$count, label_list = OSCC$group, groups = c('normal-tumor'))

}
\author{
Erjie Zhao \href{mailto:2055469819@qq.com}{2055469819@qq.com}
}
