% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{diff_limma_count}
\alias{diff_limma_count}
\title{Do differential expression analysis on read count data using limma}
\usage{
diff_limma_count(count_mat, label_list, groups, methods = c("voom", "limma-trend"))
}
\arguments{
\item{count_mat}{A matrix-like data object containing the read count data, with rows corresponding to genes and columns to samples.}

\item{label_list}{a dataframe contain samples in first column and groups information
in second column in the form of factors.}

\item{groups}{expressions, or character strings which can be parsed to expressions,
specifying contrasts, it corresponds to \code{\link[limma]{makeContrasts}}.}

\item{methods}{The method to the differential analysis. \code{voom} is powerful when
the library sizes are quite variable between samples, \code{limma-trend} the simplest
and most robust approach to differential exis when the sequencing depth is reasonably
consistent across the RNA samples.}
}
\value{
A dataframe with a row for the number top genes and the following columns:
\item{logFC}{estimate of the log2-fold-change corresponding to the effect or contrast}
\item{AveExpr}{average log2-expression for the probe over all arrays and channels, same as Amean in the \code{MarrayLM} object}
\item{t}{moderated t-statistic (omitted for \code{topTableF})}
\item{P.Value}{raw p-value}
\item{adj.P.Value}{adjusted p-value or q-value}
\item{B}{log-odds that the gene is differentially expressed (omitted for \code{topTreat})}
}
\description{
Do differential expression analysis using \code{limma} package,
this is for the read count data in RNA-seq.
}
\examples{
data(OSCC)
res <- diff_limma_count(count_mat = OSCC$count, label_list = OSCC$group,
                        groups = c("normal-tumor"), methods = "voom")

}
\author{
Erjie Zhao \href{mailto:2055469819@qq.com}{2055469819@qq.com}
}
