% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eset}
\alias{eset}
\title{microarray expression data}
\format{
\describe{
\item{array}{array probe profile}
\item{group}{sample group}
}
}
\usage{
eset
}
\description{
A list contain a microarray probe profile and a sample groups
}
\keyword{datasets}
