% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OSCC}
\alias{OSCC}
\title{RNA-seq expression data}
\format{
\describe{
\item{count}{read count matrix}
\item{group}{sample group}
\item{rpkm}{normalize rpkm matrix}
}
}
\usage{
OSCC
}
\description{
A list contain raw read count data, normalize rpkm data, and sample groups
}
\keyword{datasets}
