% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma.R
\name{diff_limma_array}
\alias{diff_limma_array}
\title{Do differential expression analysis on array data using limma}
\usage{
diff_limma_array(array_mat, label_list, groups)
}
\arguments{
\item{array_mat}{A matrix-like data object containing log-ratios or log-expression
values for a series of arrays, with rows corresponding to genes and columns to samples.}

\item{label_list}{a dataframe contain samples in first column and groups information
in second column in the form of factors.}

\item{groups}{expressions, or character strings which can be parsed to expressions,
specifying contrasts, it corresponds to \code{\link[limma]{makeContrasts}}.}
}
\value{
A dataframe with a row for the number top genes and the following columns:
\item{logFC}{estimate of the log2-fold-change corresponding to the effect or contrast}
\item{AveExpr}{average log2-expression for the probe over all arrays and channels, same as Amean in the \code{MarrayLM} object}
\item{t}{moderated t-statistic (omitted for \code{topTableF})}
\item{P.Value}{raw p-value}
\item{adj.P.Value}{adjusted p-value or q-value}
\item{B}{log-odds that the gene is differentially expressed (omitted for \code{topTreat})}
}
\description{
Do differential expression analysis using \code{limma} package,
this is for the microarray data.
}
\examples{
data(eset)
res <- diff_limma_array(array_mat = eset$array, label_list = eset$group, groups = c("brain-liver"))
}
\author{
Erjie Zhao \href{mailto:2055469819@qq.com}{2055469819@qq.com}
}
